ipq40xx: tp-link deco-m5: Split up v1 & v3 DTS
For consistency with v2, the combined v1/v3 dts has been split into two
separate variants. This is also because the led configuration of the v1
model is different than the v3 model, though as of right now the correct
led configuration isn't known yet.

As an additional benefit, once the right firmware for the specific device
variant has been flashed, there will be an image check error if an attempt
is made to flash the wrong firmware.

Signed-off-by: Bee Cadorna <r3usrlnd@gmail.com>